Move RelevantGuideIcons component to its own component folder
