+++ content/webapp/pages/guides/exhibitions/[id]/[type].tsx
@@ -50,12 +50,7 @@ import { components } from '@weco/common/views/slices';
 import { Container } from '@weco/common/views/components/styled/Container';
 import PageHeader from '@weco/common/views/components/PageHeader/PageHeader';
 import ConditionalWrapper from '@weco/common/views/components/ConditionalWrapper/ConditionalWrapper';
-import Icon from '@weco/common/views/components/Icon/Icon';
-import {
-  audioDescribed,
-  britishSignLanguage,
-  speechToText,
-} from '@weco/common/icons';
+import RelevantGuideIcons from '@weco/content/components/ExhibitionGuideRelevantIcons';
 
 const ButtonWrapper = styled(Space).attrs({
   $v: { size: 's', properties: ['margin-bottom'] },
@@ -103,50 +98,6 @@ function getTypeTitle(type: ExhibitionGuideType, egWork?: boolean): string {
       return 'Captions and transcripts';
   }
 }
-
-export const RelevantIcons = ({ types }: { types: ExhibitionGuideType[] }) => {
-  // The captions icon will be on every Guide moving forward
-  // We're ordering icons alphabetically
-  const sortedTypes = [
-    ...new Set([...types, 'captions-and-transcripts'].sort()),
-  ];
-
-  return (
-    <>
-      {sortedTypes.map((type, i) => {
-        const getIcon = () => {
-          switch (type) {
-            case 'bsl':
-              return britishSignLanguage;
-            case 'audio-without-descriptions':
-              return audioDescribed;
-            case 'captions-and-transcripts':
-              return speechToText;
-          }
-        };
-
-        const icon = getIcon();
-
-        return icon ? (
-          <ConditionalWrapper
-            key={type}
-            condition={i > 0}
-            wrapper={children => (
-              <Space
-                $h={{ size: 's', properties: ['margin-left'] }}
-                style={{ display: 'inline' }}
-              >
-                {children}
-              </Space>
-            )}
-          >
-            <Icon icon={icon} sizeOverride="height: 32px; width: 32px;" />
-          </ConditionalWrapper>
-        ) : undefined;
-      })}
-    </>
-  );
-};
 
 type Props = {
   exhibitionGuide: ExhibitionGuide | ExhibitionText | ExhibitionHighlightTour;
@@ -406,7 +357,7 @@ const ExhibitionGuidePage: FunctionComponent<Props> = props => {
           )}
 
           {egWork ? (
-            <RelevantIcons types={[type]} />
+            <RelevantGuideIcons types={[type]} />
           ) : (
             <>
               <ButtonWrapper>
